justifying center when no filter available
